Move developer commands under dev category
