+++ src/cli/commands/dev/eval-fixture.ts
@@ -1,0 +1,3 @@
+// eval-fixture command: extracted from main.ts inline action body
+// This file exists to hold the runEvalFixture logic; it is imported by categories/dev.ts
+export type { };
