+++ src/cli/commands/dev/tune.ts
@@ -1,0 +1,3 @@
+// tune command: extracted from main.ts inline action body
+// This file exists to hold the runTune logic; it is imported by categories/dev.ts
+export type { };
