+++ tests/commands/self-eval.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, vi } from "vitest";
 
 vi.setConfig({ testTimeout: 30000 });
 import { DuckDBStore } from "../../src/cli/storage/duckdb.js";
-import { runSelfEval, getScenarios, formatSelfEvalReport } from "../../src/cli/commands/self-eval.js";
+import { runSelfEval, getScenarios, formatSelfEvalReport } from "../../src/cli/commands/dev/self-eval.js";
 
 const createStore = async () => {
   const s = new DuckDBStore(":memory:");
