+++ tests/pipeline/accumulation.test.ts
@@ -3,7 +3,7 @@ import { DuckDBStore } from "../../src/cli/storage/duckdb.js";
 import { loadCore } from "../../src/cli/core/loader.js";
 import { loadFixtureIntoStore } from "../../src/cli/eval/fixture-loader.js";
 import { analyzeProject, recommendSampling } from "../../src/cli/commands/collect/calibrate.js";
-import { trainModel } from "../../src/cli/commands/train.js";
+import { trainModel } from "../../src/cli/commands/dev/train.js";
 import { planSample } from "../../src/cli/commands/sample.js";
 import { runInsights } from "../../src/cli/commands/analyze/insights.js";
 import { mkdirSync, rmSync } from "node:fs";
